Network.IRC.Client.Types

Description

Types for IRC clients. See also Network.IRC.Conduit and Network.IRC.CTCP.

Type synonyms

type UnicodeEvent = Event Text

type UnicodeSource = Source Text

type UnicodeMessage = Message Text

State

type IRC a = StatefulIRC () a

The IRC monad.

type StatefulIRC s a = ReaderT (IRCState s) IO a

The IRC monad, with state.

data IRCState s

Constructors

IRCState

Fields

_connectionConfig :: ConnectionConfig s

Read-only connection configuration

_userState :: TVar s

Mutable user state

_instanceConfig :: TVar (InstanceConfig s)

Mutable instance configuration in STM

newIRCState :: MonadIO m => ConnectionConfig s -> InstanceConfig s -> s -> m (IRCState s)

Construct a new IRC state

ircState :: StatefulIRC s (IRCState s)

Access the client state.

getConnectionConfig :: IRCState s -> ConnectionConfig s

Extract the connection configuration from an IRC state

getInstanceConfig :: IRCState s -> TVar (InstanceConfig s)

Extract the instance configuration from an IRC state

getUserState :: IRCState s -> TVar s

Extract the user state from an IRC state

getInstanceConfig' :: MonadIO m => IRCState s -> m (InstanceConfig s)

Extract the current snapshot of the instance configuration from an IRC state

connectionConfig :: StatefulIRC s (ConnectionConfig s)

Access the connection config

instanceConfigTVar :: StatefulIRC s (TVar (InstanceConfig s))

Access the instance config TVar

instanceConfig :: StatefulIRC s (InstanceConfig s)

Access the instance config as it is right now.

putInstanceConfig :: InstanceConfig s -> StatefulIRC s ()

Overwrite the instance config, even if it has changed since we looked at it.

stateTVar :: StatefulIRC s (TVar s)

Access the user state.

state :: StatefulIRC s s

Access the user state as it is right now.

putState :: s -> StatefulIRC s ()

Set the user state.

data Origin

The origin of a message.

Constructors

FromServer

FromClient

data ConnectionConfig s

The static state of an IRC server connection.

Constructors

ConnectionConfig

Fields

_func :: Int -> ByteString -> IO () -> Consumer (Either ByteString IrcEvent) IO () -> Producer IO IrcMessage -> IO ()

Function to connect and start the conduits.

_sendqueue :: TBMChan IrcMessage

Message send queue.

_server :: ByteString

The server host.

_port :: Int

The server port.

_flood :: NominalDiffTime

The minimum time between two adjacent messages.

_disconnect :: StatefulIRC s ()

Action to run if the remote server closes the connection.

_logfunc :: Origin -> ByteString -> IO ()

Function to log messages sent to and received from the server.

data InstanceConfig s

The updateable state of an IRC connection.

Constructors

InstanceConfig

Fields

_nick :: Text

Client nick

_username :: Text

Client username

_realname :: Text

Client realname

_channels :: [Text]

Current channels

_ctcpVer :: Text

Response to CTCP VERSION

_eventHandlers :: [EventHandler s]

The registered event handlers

_ignore :: [(Text, Maybe Text)]

List of nicks (optionally restricted to channels) to ignore messages from. No channel = global.

Events

data EventType

Types of events which can be caught.

Constructors

EEverything
Match all events

ENothing
Match no events

EPrivmsg

ENotice

ECTCP

ENick

EJoin

EPart

EQuit

EMode

ETopic

EInvite

EKick

EPing

ENumeric

data EventHandler s

A function which handles an event.

Constructors

EventHandler

Fields

_description :: Text

A description of the event handler.

_matchType :: EventType

Which type to be triggered by

_eventFunc :: UnicodeEvent -> StatefulIRC s ()

The function to call.

eventType :: Event a -> EventType

Get the type of an event.

Re-exported

data Event a :: * -> *

Constructors

Event

Fields

_raw :: ByteString

_source :: Source a

_message :: Message a

data Source a :: * -> *

Constructors

User (NickName a)

Channel (ChannelName a) (NickName a)

Server (ServerName a)

data Message a :: * -> *

Constructors

Privmsg (Target a) (Either CTCPByteString a)

Notice (Target a) (Either CTCPByteString a)

Nick (NickName a)

Join (ChannelName a)

Part (ChannelName a) (Reason a)

Quit (Reason a)

Mode (Target a) IsModeSet [ModeFlag a] [ModeArg a]

Topic (ChannelName a) a

Invite (ChannelName a) (NickName a)

Kick (ChannelName a) (NickName a) (Reason a)

Ping (ServerName a) (Maybe (ServerName a))

Pong (ServerName a)

Numeric Int [NumericArg a]

RawMsg a