Network.IRC.Client.Utils

Description

Commonly-used utility functions for IRC clients.

Documentation

setNick :: Text -> IRC ()

Update the nick in the instance configuration and also send an update message to the server.

leaveChannel :: Text -> Maybe Text -> IRC ()

Update the channel list in the instance configuration and also part the channel.

delChan :: TVar InstanceConfig -> Text -> STM ()

Remove a channel from the list.

reply :: UnicodeEvent -> Text -> IRC ()

Send a message to the source of an event.

ctcp :: Text -> Text -> [Text] -> UnicodeMessage

Construct a privmsg containing a CTCP

ctcpReply :: Text -> Text -> [Text] -> UnicodeMessage

Construct a notice containing a CTCP